pyproject.toml
[tool.poetry]
name = "gcv"
version = "0.1.0"
description = "Gestão Centralizada de Vulnerabilidades"
authors = ["google-labs-jules[bot] <161369871+google-labs-jules[bot]@users.noreply.github.com>"]
packages = [
    { include = "app", from = "gcv/backend/src" },
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-multipart = "^0.0.7"
pydantic = "^2.5.3"
psycopg2-binary = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyotp = "^2.8.0"
google-generativeai = "^0.4.1"
python-sonarqube-api = "*"
authlib = "^1.2.1"
apscheduler = "^3.10.4"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
sqlalchemy = "^2.0.23"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
